Answer an annotated use of a variable with the use itself, and drop the delegate
The delegating element that carried the annotations of a use onto the type the
variable is bound to is gone, along with the copy of it the Python processor
had to extend. A delegate has to be kept in step with every default method the
interface gains, and it bought nothing the generated metadata keeps: the writers
record no variable name for an annotated type argument at all, and they resolve
a variable before reading its annotations, so a variable bound to a concrete
type never carried the annotations of the use, on 5.2.x either.

An annotated use is now answered with the element of the use, which already
stands for what the variable is bound to and is the only element carrying those
annotations. The variable name it reports is the one the intermediate type
writes it with, which is what the generated argument says for an annotated type
argument in any case. Everything unannotated binds as before, Kotlin included,
where the use carries no resolution of its own.

`withAnnotationMetadata` cannot do this: it presets the annotation metadata of
the element, while the writers persist `getTypeAnnotationMetadata`, which no
implementation reads the preset for.
